Extract WooPlanCardView to its own file
