+++ WooCommerce/Classes/ViewRelated/Upgrades/WooPlanCardView.swift
@@ -1,0 +1,142 @@
+import Yosemite
+import SwiftUI
+
+struct WooPlanCardView: View {
+    let upgradePlan: WooWPComPlan
+    @Binding var selectedPlan: WooWPComPlan?
+
+    private var isSelected: Bool {
+        selectedPlan?.id == upgradePlan.id
+    }
+
+    private var isPopular: Bool {
+        let popularPlans =  [
+            AvailableInAppPurchasesWPComPlans.performanceMonthly.rawValue,
+            AvailableInAppPurchasesWPComPlans.performanceYearly.rawValue
+        ]
+
+        return popularPlans.contains(where: {$0 == upgradePlan.id})
+    }
+    @State private var isExpanded = false
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: Layout.spacing) {
+            VStack(alignment: .leading, spacing: Layout.spacing) {
+                HStack {
+                    Text(upgradePlan.wooPlan.shortName)
+                        .font(.title2)
+                        .bold()
+                        .accessibilityAddTraits(.isHeader)
+
+                    Spacer()
+
+                    BadgeView(text: Localization.isPopularBadgeText.uppercased()).renderedIf(isPopular)
+                    Image(systemName: isSelected ? "checkmark.circle.fill" : "circle")
+                        .foregroundStyle(isSelected ? Color.withColorStudio(name: .wooCommercePurple, shade: .shade50) : Color(.systemGray4))
+                        .font(.system(size: Layout.checkImageSize))
+
+                }
+                Text(upgradePlan.wooPlan.planDescription)
+                    .font(.subheadline)
+            }
+
+            VStack(alignment: .leading, spacing: Layout.textSpacing) {
+                Text(upgradePlan.wpComPlan.displayPrice)
+                    .font(.title)
+                    .bold()
+                    .accessibilityAddTraits(.isHeader)
+                Text(upgradePlan.wooPlan.planFrequency.localizedString)
+                    .font(.footnote)
+            }
+
+            let buttonText = String.localizedStringWithFormat(Localization.viewPlanFeaturesFormat, upgradePlan.wooPlan.shortName)
+            Button("\(buttonText) \(Image(systemName: isExpanded ? "chevron.up" : "chevron.down"))") {
+                isExpanded.toggle()
+            }
+            WooPlanCardFeaturesView(upgradePlan.wooPlan.shortName).renderedIf(isExpanded)
+        }
+        .frame(maxWidth: .infinity)
+        .padding(.horizontal)
+        .padding(.vertical)
+        .background(Color(.systemGroupedBackground))
+        .cornerRadius(Layout.cornerRadius)
+        .overlay(
+            RoundedRectangle(cornerRadius: Layout.cornerRadius)
+                .stroke(isSelected ? .withColorStudio(name: .wooCommercePurple, shade: .shade50) : Color(.systemGray4),
+                        lineWidth: isSelected ? Layout.selectedBorder : Layout.unselectedBorder)
+        )
+        .contentShape(Rectangle())
+        .onTapGesture {
+            if selectedPlan?.id != upgradePlan.id {
+                selectedPlan = upgradePlan
+            }
+        }
+    }
+}
+
+private struct WooPlanCardFeaturesView: View {
+
+    private let selectedPlan: String
+
+    init(_ selectedPlan: String) {
+        self.selectedPlan = selectedPlan
+    }
+
+    var planFeatures: [String] {
+        WooPlan.loadHardcodedPlan(selectedPlan).planFeatures
+    }
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: Layout.featureTextSpacing) {
+            Text(Localization.upsellFeatureTitleText)
+                .bold()
+                .font(.footnote)
+                .renderedIf(selectedPlan != "Essential")
+            ForEach(planFeatures, id: \.self) { feature in
+                Text(feature)
+                    .font(.footnote)
+                    .multilineTextAlignment(.leading)
+            }
+            Text(Localization.storageText.uppercased())
+                .bold()
+                .font(.footnote)
+            BadgeView(text: Localization.storageAmountText, customizations: .init(textColor: Color(.text), backgroundColor: Color(.systemGray4)))
+                .font(.footnote)
+        }
+    }
+}
+
+private extension WooPlanCardFeaturesView {
+    enum Localization {
+        static let upsellFeatureTitleText = NSLocalizedString("Everything in Essential, plus:", comment: "")
+        static let storageText = NSLocalizedString("Storage", comment: "")
+        static let storageAmountText = NSLocalizedString("50 GB", comment: "")
+    }
+
+    enum Layout {
+        static let featureTextSpacing: CGFloat = 8.0
+    }
+}
+
+private extension WooPlanCardView {
+    enum Layout {
+        static let cornerRadius: CGFloat = 8.0
+        static let selectedBorder: CGFloat = 2
+        static let unselectedBorder: CGFloat = 0.5
+        static let checkImageSize: CGFloat = 24
+        static let spacing: CGFloat = 16
+        static let textSpacing: CGFloat = 4
+    }
+
+    enum Localization {
+        static let viewPlanFeaturesFormat = NSLocalizedString(
+            "View %1$@ features",
+            comment: "Title for the button to expand plan details on the Upgrade plan screen. " +
+            "Reads as 'View Essential features'. %1$@ must be included in the string and will be replaced with " +
+            "the plan name.")
+
+        static let isPopularBadgeText = NSLocalizedString(
+            "Popular",
+            comment: "The text of the badge that indicates the most popular choice when purchasing a Plan")
+    }
+}
